Try assets instead of /assets
The project page for this repo is nickduckwiler.com/stretch-dreams. Thus
I assume the assets should be located at nickduckwiler.com/stretch-dreams/assets.

As I am writing the hrefs link to "/assets/..."

Linking to "/assets/..." is absolute, so the browser interprets that to be
nickduckwiler.com/assets/....

Linking to "assets/..." is relative, so the browser interprets that to
be nickduckwiler.com/stretch-dreams/assets/..., which, I believe, is
what I want